task_management/user-service/pyproject.toml
[project]
name = "fast-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.122.0",
    "fastapi-limiter>=0.1.6",
    "loguru>=0.7.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-requests",
    "opentelemetry-instrumentation-sqlalchemy",
    "aiokafka>=0.12.0",
]
